requirements.txt
# TrustCoin Bot Dependencies
# Core Telegram Bot Library
python-telegram-bot==20.7

# Environment Variables Management
python-dotenv==1.0.0

# HTTP requests and async support
requests==2.31.0
aiohttp==3.9.1

# Web framework for webhooks
Flask==3.0.0

# Process monitoring and management
psutil==5.9.6

# For production deployment
gunicorn==21.2.0
uvicorn==0.24.0

# Logging and monitoring
loguru==0.7.2

# Security enhancements
cryptography==41.0.7
bcrypt==4.1.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Additional utilities
colorama==0.4.6
rich==13.7.0